# api/openapi_chatgpt_operations_v016.yaml
openapi: 3.1.0
info:
  title: SPPG Vendor and Inventory Operations
  version: 0.16.0
  description: Parse vendor WhatsApp invoices, reconcile PO/receiving/invoice/reject layers, manage operational stock, and confirm vendor payment evidence. No finance ledger transaction is created automatically by these operations.
servers:
  - url: https://sppg-finance-gpt-site-production-5b7d.up.railway.app
security:
  - bearerAuth: []
paths:
  /v1/vendor-invoices/parse-whatsapp:
    post:
      operationId: parseSppgVendorWhatsAppInvoice
      summary: Parse vendor invoice text and generate a payment draft
      description: Read-only parser. Extracts invoice lines, validates arithmetic, applies reject deductions, and generates WhatsApp-ready payment text. Does not write database records.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VendorInvoiceTextRequest'
      responses:
        '200':
          description: Parsed invoice and payment draft
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorInvoiceParseResponse'
  /v1/vendor-payables/from-receipt:
    post:
      operationId: processSppgVendorPayableFromReceipt
      summary: Preview or commit reconciled vendor payable
      description: Always use commit=false first. Keep PO qty, received qty, invoice qty, rejected qty, and payable qty separate. vendor_cost_price must be explicit.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VendorPayableRequest'
      responses:
        '200':
          description: Payable preview or commit result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorPayableResponse'
        '409':
          description: Item or quantity reconciliation is ambiguous
  /v1/vendor-payables:
    get:
      operationId: searchSppgVendorPayables
      summary: Search vendor payables
      parameters:
        - in: query
          name: site
          schema: {type: string, enum: [MAJA, CEMPLANG]}
        - in: query
          name: vendor
          schema: {type: string}
        - in: query
          name: status
          schema: {type: string, enum: [UNPAID, PARTIAL, PAID, CANCELLED]}
        - in: query
          name: limit
          schema: {type: integer, minimum: 1, maximum: 500, default: 200}
      responses:
        '200':
          description: Matching payables
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayableSearchResponse'
  /v1/vendor-payments/confirm:
    post:
      operationId: confirmSppgVendorPayment
      summary: Preview or confirm vendor payment evidence
      description: Use commit=false first. On commit, records payment and updates payable status to PARTIAL or PAID. Does not create a finance ledger transaction automatically.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VendorPaymentRequest'
      responses:
        '200':
          description: Payment preview or confirmation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VendorPaymentResponse'
        '409':
          description: Payment exceeds remaining payable
  /v1/inventory/from-receipt:
    post:
      operationId: postSppgReceiptToInventory
      summary: Preview or post confirmed receipt into operational stock
      description: Idempotent. Creates stock movements only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReceiptToStockRequest'
      responses:
        '200':
          description: Inventory receipt result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptToStockResponse'
  /v1/inventory/usage:
    post:
      operationId: processSppgInventoryUsage
      summary: Preview or record actual stock usage
      description: Use commit=false first. Operational stock only, no finance transaction.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryUsageRequest'
      responses:
        '200':
          description: Usage result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryUsageResponse'
  /v1/inventory/balance:
    get:
      operationId: getSppgInventoryBalance
      summary: Read current stock balance for one item
      parameters:
        - in: query
          name: site
          required: true
          schema: {type: string, enum: [MAJA, CEMPLANG]}
        - in: query
          name: item
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Current stock balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryBalanceResponse'
  /v1/inventory/requirement-preview:
    get:
      operationId: previewSppgInventoryRequirement
      summary: Calculate purchase requirement after current stock
      description: Useful for Wikian chicken ordered for multiple production days.
      parameters:
        - in: query
          name: site
          required: true
          schema: {type: string, enum: [MAJA, CEMPLANG]}
        - in: query
          name: item
          required: true
          schema: {type: string}
        - in: query
          name: plannedQty
          required: true
          schema: {type: number, minimum: 0}
        - in: query
          name: unit
          schema: {type: string, default: kg}
      responses:
        '200':
          description: Stock-adjusted purchase requirement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryRequirementResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    VendorInvoiceTextRequest:
      type: object
      required: [site, vendor_code, invoice_date_label, text]
      properties:
        site: {type: string, enum: [MAJA, CEMPLANG]}
        vendor_code: {type: string}
        invoice_date_label: {type: string}
        text: {type: string}
    ParsedInvoiceLine:
      type: object
      properties:
        reported_item_name: {type: string}
        item_name: {type: string}
        invoiced_qty: {type: number}
        unit: {type: string}
        vendor_cost_price: {type: number}
        declared_line_total: {type: number}
        computed_line_total: {type: number}
        line_total_matches: {type: boolean}
        rejected_qty: {type: number}
        reject_amount: {type: number}
        payable_qty: {type: number}
        net_line_total: {type: number}
        reject_match_confidence: {type: number}
    VendorInvoiceParseResponse:
      type: object
      properties:
        vendorCode: {type: string}
        site: {type: string}
        declaredTotal: {type: number}
        grossAmount: {type: number}
        rejectDeduction: {type: number}
        netAmount: {type: number}
        canCommit: {type: boolean}
        financeTransactionCreated: {type: boolean}
        paymentDraft: {type: string}
        warnings:
          type: array
          items: {type: string}
        items:
          type: array
          items:
            $ref: '#/components/schemas/ParsedInvoiceLine'
    VendorCostLine:
      type: object
      required: [vendor_cost_price]
      properties:
        goods_receipt_item_id: {type: [integer, 'null']}
        item_name: {type: [string, 'null']}
        vendor_cost_price: {type: number, minimum: 0}
        invoiced_qty: {type: [number, 'null'], minimum: 0}
        rejected_qty: {type: number, minimum: 0, default: 0}
    VendorPayableRequest:
      type: object
      required: [site, purchase_order_id, goods_receipt_id, commit, lines]
      properties:
        site: {type: string, enum: [MAJA, CEMPLANG]}
        purchase_order_id: {type: integer}
        goods_receipt_id: {type: integer}
        invoice_number: {type: [string, 'null']}
        invoice_date: {type: [string, 'null'], format: date}
        due_date: {type: [string, 'null'], format: date}
        evidence_uri: {type: [string, 'null']}
        commit: {type: boolean, default: false}
        lines:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/VendorCostLine'
    VendorPayableLine:
      type: object
      properties:
        goods_receipt_item_id: {type: integer}
        purchase_order_item_id: {type: [integer, 'null']}
        item_name: {type: string}
        accepted_qty: {type: number}
        invoiced_qty: {type: number}
        rejected_qty: {type: number}
        payable_qty: {type: number}
        unit: {type: [string, 'null']}
        planned_qty: {type: [number, 'null']}
        po_qty: {type: [number, 'null']}
        planning_price: {type: [number, 'null']}
        po_price: {type: [number, 'null']}
        vendor_cost_price: {type: number}
        gross_line_total: {type: number}
        reject_amount: {type: number}
        line_total: {type: number}
        invoice_vs_po_variance: {type: [number, 'null']}
        invoice_vs_receipt_variance: {type: number}
        warnings:
          type: array
          items: {type: string}
    VendorPayableResponse:
      type: object
      properties:
        committed: {type: boolean}
        canCommit: {type: boolean}
        duplicate: {type: boolean}
        vendorInvoiceId: {type: [integer, 'null']}
        site: {type: string}
        purchaseOrderId: {type: integer}
        poCode: {type: string}
        goodsReceiptId: {type: integer}
        vendorCode: {type: string}
        payableStatus: {type: string}
        grossAmount: {type: number}
        rejectDeduction: {type: number}
        netAmount: {type: number}
        financeTransactionCreated: {type: boolean}
        warnings:
          type: array
          items: {type: string}
        lines:
          type: array
          items:
            $ref: '#/components/schemas/VendorPayableLine'
    PayableRecord:
      type: object
      properties:
        vendor_invoice_id: {type: integer}
        vendor_code: {type: string}
        site: {type: string}
        purchase_order_id: {type: [integer, 'null']}
        goods_receipt_id: {type: [integer, 'null']}
        invoice_number: {type: [string, 'null']}
        gross_amount: {type: number}
        reject_deduction: {type: number}
        net_amount: {type: number}
        payable_status: {type: string}
    PayableSearchResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PayableRecord'
    VendorPaymentRequest:
      type: object
      required: [vendor_invoice_id, amount, commit]
      properties:
        vendor_invoice_id: {type: integer}
        amount: {type: number, exclusiveMinimum: 0}
        paid_at: {type: [string, 'null'], format: date-time}
        payment_source: {type: [string, 'null']}
        reference_number: {type: [string, 'null']}
        evidence_uri: {type: [string, 'null']}
        source_external_id: {type: [string, 'null']}
        commit: {type: boolean, default: false}
    VendorPaymentResponse:
      type: object
      properties:
        committed: {type: boolean}
        duplicate: {type: boolean}
        vendorPaymentId: {type: [integer, 'null']}
        vendorInvoiceId: {type: integer}
        vendorCode: {type: string}
        site: {type: string}
        invoiceNumber: {type: [string, 'null']}
        netAmount: {type: number}
        alreadyPaid: {type: number}
        paymentAmount: {type: number}
        remainingBefore: {type: number}
        remainingAfter: {type: number}
        payableStatusAfter: {type: string}
        canCommit: {type: boolean}
        financeTransactionCreated: {type: boolean}
    ReceiptToStockRequest:
      type: object
      required: [site, goods_receipt_id, commit]
      properties:
        site: {type: string, enum: [MAJA, CEMPLANG]}
        goods_receipt_id: {type: integer}
        commit: {type: boolean, default: false}
    StockMovementPreview:
      type: object
      properties:
        sourceKey: {type: string}
        itemName: {type: string}
        qty: {type: number}
        unit: {type: [string, 'null']}
        fromLocation: {type: string}
        toLocation: {type: string}
    ReceiptToStockResponse:
      type: object
      properties:
        committed: {type: boolean}
        canCommit: {type: boolean}
        goodsReceiptId: {type: integer}
        inserted: {type: integer}
        duplicates: {type: integer}
        items:
          type: array
          items:
            $ref: '#/components/schemas/StockMovementPreview'
    InventoryUsageRequest:
      type: object
      required: [site, item_name, qty, unit, commit]
      properties:
        site: {type: string, enum: [MAJA, CEMPLANG]}
        item_name: {type: string}
        qty: {type: number, exclusiveMinimum: 0}
        unit: {type: string}
        occurred_at: {type: [string, 'null'], format: date-time}
        source_ref: {type: [string, 'null']}
        commit: {type: boolean, default: false}
    InventoryUsageResponse:
      type: object
      properties:
        committed: {type: boolean}
        duplicate: {type: boolean}
        movementId: {type: [integer, 'null']}
        site: {type: string}
        itemName: {type: string}
        balanceBefore: {type: number}
        usageQty: {type: number}
        balanceAfter: {type: number}
        unit: {type: string}
        stockWarning: {type: boolean}
    InventoryBalanceResponse:
      type: object
      properties:
        site: {type: string}
        itemName: {type: string}
        balance: {type: number}
    InventoryRequirementResponse:
      type: object
      properties:
        site: {type: string}
        itemName: {type: string}
        plannedQty: {type: number}
        stockAvailable: {type: number}
        purchaseNeeded: {type: number}
        unit: {type: string}
        financeTransactionCreated: {type: boolean}